+++ Changelog.md
@@ -4,6 +4,7 @@
 * Add check of droplet size for creation
 * Do not force 2.0.1 of `droplet_kit`
 * Change default region to `nyc3`
+* Add ability to retry request if error happened
 
 # 0.2.0
 * Force use ruby 2.2 to be able use `activesupport 5`